add authorization

+++ planner/routes/events.py
@@ -1,7 +1,8 @@
-from fastapi import APIRouter, HTTPException, status, Body
+from fastapi import APIRouter, HTTPException, status, Body, Depends
 from planner.models.events import Event, EventUpdate
 from beanie import PydanticObjectId
 from planner.database.connection import Database
+from auth.authenticate import authenticate
 
 event_router = APIRouter(tags=["Event"])
 
@@ -26,7 +27,8 @@ async def retrieve_event(id: PydanticObjectId) -> Event:
 
 
 @event_router.post("/")
-async def create_event(body: Event) -> dict:
+async def create_event(body: Event, user: str = Depends(authenticate)) -> dict:
+    body.creator = user
     await event_database.save(body)
     return {
         "message": "Event created successfully"
@@ -34,7 +36,13 @@ async def create_event(body: Event) -> dict:
 
 
 @event_router.put('/{id}', response_model=Event)
-async def update_event(id: PydanticObjectId, body: EventUpdate) -> Event:
+async def update_event(id: PydanticObjectId, body: EventUpdate, user: str = Depends(authenticate)) -> Event:
+    event = await event_database.get(id)
+    if event.creator != user:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="Operation not allowed"
+        )
     updated_event = await event_database.update(id, body)
     if not updated_event:
         raise HTTPException(
@@ -46,7 +54,13 @@ async def update_event(id: PydanticObjectId, body: EventUpdate) -> Event:
 
 
 @event_router.delete("/{id}")
-async def delete_event(id: PydanticObjectId) -> bool:
+async def delete_event(id: PydanticObjectId, user: str = Depends(authenticate)) -> bool:
+    event = await event_database.get(id)
+    if event.creator != user:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="Operation not allowed"
+        )
     event = await event_database.get(id)
     if event:
         return await event_database.delete(id)
